refactor(executions): extract the concurrency slot release from DefaultExecutor into a processor
